Refactoring: move logstash to logging/logstash
