+++ src/main/java/tech/jhipster/lite/generator/server/springboot/logging/logstash/domain/LogstashDomainService.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.server.springboot.logstash.tcp.domain;
+package tech.jhipster.lite.generator.server.springboot.logging.logstash.domain;
 
 import static tech.jhipster.lite.common.domain.FileUtils.getPath;
 import static tech.jhipster.lite.generator.project.domain.Constants.MAIN_JAVA;
