+++ src/main/java/tech/jhipster/lite/generator/server/springboot/logging/logstash/domain/LogstashService.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.server.springboot.logstash.tcp.domain;
+package tech.jhipster.lite.generator.server.springboot.logging.logstash.domain;
 
 import tech.jhipster.lite.generator.project.domain.Project;
