+++ src/test/java/tech/jhipster/lite/generator/server/springboot/logging/logstash/infrastructure/config/LogstashBeanConfigurationIT.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.server.springboot.logstash.tcp.infrastructure.config;
+package tech.jhipster.lite.generator.server.springboot.logging.logstash.infrastructure.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import tech.jhipster.lite.IntegrationTest;
-import tech.jhipster.lite.generator.server.springboot.logstash.tcp.domain.LogstashDomainService;
+import tech.jhipster.lite.generator.server.springboot.logging.logstash.domain.LogstashDomainService;
 
 @IntegrationTest
 class LogstashBeanConfigurationIT {
