+++ tests-ci/generate.sh
@@ -35,7 +35,7 @@ if [[ $filename == 'tomcat-postgresql-ehcachejava' ]]; then
   callApi "/api/servers/spring-boot/databases/postgresql"
   callApi "/api/servers/spring-boot/databases/migration/liquibase"
   callApi "/api/servers/spring-boot/logging/aop"
-  callApi "/api/servers/spring-boot/logging/logstash/tcp"
+  callApi "/api/servers/spring-boot/logging/logstash"
   callApi "/api/servers/spring-boot/async"
   callApi "/api/servers/spring-boot/cache/ehcache/java-configuration"
   callApi "/api/servers/sonar"
@@ -55,7 +55,7 @@ elif [[ $filename == 'tomcat-postgresql-ehcachexml' ]]; then
   callApi "/api/servers/spring-boot/databases/postgresql"
   callApi "/api/servers/spring-boot/databases/migration/liquibase"
   callApi "/api/servers/spring-boot/logging/aop"
-  callApi "/api/servers/spring-boot/logging/logstash/tcp"
+  callApi "/api/servers/spring-boot/logging/logstash"
   callApi "/api/servers/spring-boot/async"
   callApi "/api/servers/spring-boot/cache/ehcache/xml-configuration"
   callApi "/api/servers/sonar"
@@ -75,7 +75,7 @@ elif [[ $filename == 'undertow-mysql-simplecache' ]]; then
   callApi "/api/servers/spring-boot/databases/mysql"
   callApi "/api/servers/spring-boot/databases/migration/liquibase"
   callApi "/api/servers/spring-boot/logging/aop"
-  callApi "/api/servers/spring-boot/logging/logstash/tcp"
+  callApi "/api/servers/spring-boot/logging/logstash"
   callApi "/api/servers/spring-boot/async"
   callApi "/api/servers/spring-boot/cache/simple"
   callApi "/api/servers/sonar"
